Update language-redirect.php
Optionally based on a config checkbox setting, tests for a cookie and if still valid then doesn't redirect. If the cookie is not set then it sets one (lifetime is 28 days) and then continues to the redirect code. Use this on sites whereby you want the root i.e. / of your multisite to be able to be browsed because that is a default language e.g. english. Not needed if all languages have their own subsites i.e. no one ever browsers around / but gets redirected to /en/, /fr/, /it/, /de/